CORS - support for x-mojo-around-action
Now CORS checks are dont prior to delegating request handling to the x-mojo-around-action hook.
This way when the browser makes a CORS-request, but the x-mojo-around-action
for some reason fails the request, the browser gets the reason of the failure, instead of the
failed CORS request hiding any responses from the server.